requirements.txt
mysql-connector-python
Jinja2
reportlab
xhtml2pdf
flask
flask-cors